controller/requirements.txt
grpcio
grpcio-tools
googleapis-common-protos
protobuf
scapy
pycryptodomex
